moved PageHeader out of pattern folder to organisms folder
