--- app/controllers/pattern_libraries_controller.rb
@@ -1,4 +1,0 @@
-class PatternLibrariesController < ApplicationController
-  def index
-  end
-end
