+++ app/controllers/pattern_library_controller.rb
@@ -1,0 +1,4 @@
+class PatternLibraryController < ApplicationController
+  def index
+  end
+end
