+++ app/javascript/components/organisms/PageHeader.js
@@ -2,7 +2,7 @@ import React from 'react'
 
 const PageHeader = ({ title, content }) => {
   return (
-    <div className='header'>
+    <div className='page-header'>
       <h1>{title}</h1>
       <p>{content}</p>
     </div>
