+++ app/javascript/packs/pattern_library.js
@@ -1,7 +1,7 @@
 import WebpackerReact from 'webpacker-react'
 
 // All imports from Pattern Library
-import PageHeader from 'components/patterns/PageHeader'
+import PageHeader from 'components/organisms/PageHeader'
 
 // All setup components after this
 WebpackerReact.setup({ PageHeader })
