+++ config/routes.rb
@@ -28,7 +28,7 @@ Rails.application.routes.draw do
     get 'marked_duplicate_index' => 'flagged_record_sets#marked_duplicate_index'
   end
 
-  resources :pattern_libraries, only: %w[index]
+  resources :pattern_library, only: %w[index]
 
   ## --- API namespacing
   namespace :api do
